SSLError workaround moved to service folder
