+++ custom_components/waste_collection_schedule/waste_collection_schedule/service/SSLError.py
@@ -1,0 +1,27 @@
+#Work around SSL UNSAFE_LEGACY_RENEGOTIATION_DISABLED errors using method discussed in
+# https://stackoverflow.com/questions/71603314/ssl-error-unsafe-legacy-renegotiation-disabled
+
+import requests
+import ssl
+import urllib3
+
+class CustomHttpAdapter (requests.adapters.HTTPAdapter):
+    # "Transport adapter" that allows us to use custom ssl_context.
+
+    def __init__(self, ssl_context=None, **kwargs):
+        self.ssl_context = ssl_context
+        super().__init__(**kwargs)
+
+    def init_poolmanager(self, connections, maxsize, block=False):
+        self.poolmanager = urllib3.poolmanager.PoolManager(
+            num_pools=connections, maxsize=maxsize,
+            block=block, ssl_context=self.ssl_context)
+
+
+def get_legacy_session():
+    ctx = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
+    ctx.options |= 0x4  # OP_LEGACY_SERVER_CONNECT
+    session = requests.session()
+    session.mount('https://', CustomHttpAdapter(ctx))
+    return session
+
