+++ custom_components/waste_collection_schedule/waste_collection_schedule/source/aucklandcouncil_govt_nz.py
@@ -1,14 +1,11 @@
 import datetime
 from html.parser import HTMLParser
 
-import requests
+# import requests
 from waste_collection_schedule import Collection  # type: ignore[attr-defined]
 
-# Updated to work around SSL UNSAFE_LEGACY_RENEGOTIATION_DISABLED error using method discussed in
-# https://stackoverflow.com/questions/71603314/ssl-error-unsafe-legacy-renegotiation-disabled
-import ssl
-import urllib3
-
+# Include work around for SSL UNSAFE_LEGACY_RENEGOTIATION_DISABLED error
+from ..service.SSLError import get_legacy_session
 
 TITLE = "Auckland council"
 DESCRIPTION = "Source for Auckland council."
@@ -32,30 +29,6 @@ MONTH = {
     "November": 11,
     "December": 12,
 }
-
-
-# Additional code snippet to work around SSL issue
-class CustomHttpAdapter (requests.adapters.HTTPAdapter):
-    # "Transport adapter" that allows us to use custom ssl_context.
-
-    def __init__(self, ssl_context=None, **kwargs):
-        self.ssl_context = ssl_context
-        super().__init__(**kwargs)
-
-    def init_poolmanager(self, connections, maxsize, block=False):
-        self.poolmanager = urllib3.poolmanager.PoolManager(
-            num_pools=connections, maxsize=maxsize,
-            block=block, ssl_context=self.ssl_context)
-
-
-def get_legacy_session():
-    ctx = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
-    ctx.options |= 0x4  # OP_LEGACY_SERVER_CONNECT
-    session = requests.session()
-    session.mount('https://', CustomHttpAdapter(ctx))
-    return session
-# End SSL issue code snippet
-
 
 
 def toDate(formattedDate):
@@ -143,13 +116,6 @@ class Source:
             params=params,
             # verify=False,
         )
-
-        # Original request code
-        # r = requests.get(
-        #     "https://www.aucklandcouncil.govt.nz/rubbish-recycling/rubbish-recycling-collections/Pages/collection-day-detail.aspx",
-        #     params=params,
-        #     verify=False,
-        # )
 
         p = WasteSearchResultsParser()
         p.feed(r.text)
